pyproject.toml
[tool.poetry]
name = "blockchain-class"
version = "0.1.0"
description = ""
authors = ["Seth Mcfeeters <seth.mcfeeters@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.4"
numpy = "^1.20.3"
matplotlib = "^3.4.2"
pycryptodome = "^3.10.1"
pytest = "^6.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
